SERVER: refactor into classes, files, linting
- classes added
- routes (blueprints) and classes into separate folders
- linting
