+++ server/project/models/constants.py
@@ -1,0 +1,9 @@
+"""
+This program has been developed by students from the bachelor Computer Science at
+Utrecht University within the Software Project course.
+© Copyright Utrecht University (Department of Information and Computing Sciences)
+"""
+
+# Constants
+RESULTS_DIR = 'results/'
+MAIL_KEY = 'MAIL'
