+++ server/project/models/events.py
@@ -11,17 +11,16 @@ from fairreckitlib.model.pipeline.model_event import ON_BEGIN_MODEL_PIPELINE, \
 from fairreckitlib.data.pipeline.data_event import ON_BEGIN_DATA_PIPELINE, \
     ON_BEGIN_FILTER_DATASET, ON_BEGIN_SPLIT_DATASET
 from fairreckitlib.core.parsing.parse_event import ON_PARSE
-from .mail import send_mail
-from .result_storage import save_result
 from .experiment import Status, ProgressStatus
 
 
-class EventHandler():
+class EventHandler:
     """Handles the events"""
 
-    def __init__(self, experiment, end_experiment):
+    def __init__(self, experiment, result_storage, mail_sender):
         self.experiment = experiment
-        self.end_experiment = end_experiment
+        self.result_storage = result_storage
+        self.mail_sender = mail_sender
         self.events = {
             # ON_BEGIN_EXPERIMENT_PIPELINE: lambda x, **kwargs: print('uwu'),
             # ON_END_EXPERIMENT_PIPELINE: self.on_end_experiment,
@@ -87,16 +86,15 @@ class EventHandler():
             if not self.experiment.validating:
                 # TODO Update experiment data: Save elapsed time
                 #self.experiment.job['metadata']['duration'] = kwargs['elapsed_time']
-                save_result(self.experiment.job, self.experiment.config)
+                self.result_storage.save_result(self.experiment.job, self.experiment.config)
                 if 'email' in self.experiment.job['metadata']:
-                    send_mail(self.experiment.job['metadata']['email'],
+                    self.mail_sender.send_mail(self.experiment.job['metadata']['email'],
                               self.experiment.job['metadata']['name'],
                               self.experiment.job['timestamp']['datetime'])
             # else: self.experiment.job['runs'] = kwargs['num_runs']
 
             self.experiment.status = Status.DONE
             self.experiment.progress = ProgressStatus.FINISHED
-        self.end_experiment()
 
 
 def do_nothing(event_listener, event_args, kwargs):
