+++ server/project/models/result_loader.py
@@ -1,4 +1,3 @@
-
 """This module contains functions to load & modify data from evaluation results.
 
 methods:
@@ -13,20 +12,21 @@ Utrecht University within the Software Project course.
 © Copyright Utrecht University (Department of Information and Computing Sciences)
 """
 import os
-#import pandas as pd
-from .result_storage import RESULTS_ROOT_FOLDER, load_results_overview, load_json
-from . import result_storage
+# import pandas as pd
+from .constants import RESULTS_DIR
+from .result_storage import load_results_overview, load_json
 
 
-def result_by_id(result_id):
+def result_by_id(result_id, result_storage):
     """Set the current result to a result in the results overview by its id.
 
     Args:
     result_id(int): the result id
+    result_store(ResultStorage): the result storage with the current result
     """
     results_overview = load_results_overview()
-    relative_path = RESULTS_ROOT_FOLDER + str(result_id) + "_" + \
-        id_to_name(results_overview, result_id)
+    relative_path = RESULTS_DIR + str(result_id) + "_" + \
+                    id_to_name(results_overview, result_id)
     data = results_overview['all_results'][id_to_index(
         results_overview, result_id)]
     data['metadata']['runs'] = 0  # Store runs
@@ -36,7 +36,7 @@ def result_by_id(result_id):
         # loops through individual results by looping through each entry in the overview.json
         for pair_id, pair_data in enumerate(run_overview['overview']):
             evaluation_path_full = os.getcwd() + "/" + \
-                pair_data['evaluation_path']
+                                   pair_data['evaluation_path']
             # ratings_settings_path_full = os.getcwd() + "/" + \
             #    pair_data['ratings_settings_path']
 
@@ -59,7 +59,29 @@ def result_by_id(result_id):
         data['metadata']['runs'] += 1
 
     result_storage.current_result = data
-    #print('current result', json.dumps(current_result, indent=4))
+    # print('current result', json.dumps(current_result, indent=4))
+
+
+def get_overview(evaluation_id, runid):
+    """Return a specific entry from a specific overview.json.
+
+    Args:
+    evaluation_id(int): the calculation id of the evaluation
+    runid(int): the id of the specific dataset-recommender approach pair
+    """
+    results_overview = load_results_overview()
+
+    name = id_to_name(results_overview, evaluation_id)
+
+    relative_path = RESULTS_DIR + str(evaluation_id) + \
+                    "_" + name + "/" + "run_" + str(runid)
+    overview_path = relative_path + "/overview.json"
+    run_overview = load_json(overview_path)
+    return run_overview['overview']
+    # rec_paths = []
+    # for index in range(0, len(run_overview['overview'])):
+    #    rec_paths[index] = run_overview['overview'][pairid]['ratings_path']
+    # return rec_paths
 
 
 def add_evaluation(data, evaluation):
@@ -84,28 +106,6 @@ def format_evaluation(evaluation):
         eval_item.pop('evaluation')
         eval_item['evaluations'] = evaluation_list
     return evaluation
-
-
-def get_overview(evaluation_id, runid):
-    """Return a specific entry from a specific overview.json.
-
-    Args:
-    evaluation_id(int): the calculation id of the evaluation
-    runid(int): the id of the specific dataset-recommender approach pair
-    """
-    results_overview = load_results_overview()
-
-    name = id_to_name(results_overview, evaluation_id)
-
-    relative_path = RESULTS_ROOT_FOLDER + str(evaluation_id) + \
-        "_" + name + "/" + "run_" + str(runid)
-    overview_path = relative_path + "/overview.json"
-    run_overview = load_json(overview_path)
-    return run_overview['overview']
-    # rec_paths = []
-    # for index in range(0, len(run_overview['overview'])):
-    #    rec_paths[index] = run_overview['overview'][pairid]['ratings_path']
-    # return rec_paths
 
 
 def id_to_name(json_data, result_id):
@@ -143,14 +143,14 @@ def name_to_index(json_data, name, key, by_name=False):
     of a specific dataset-recommender approach pair.
 
     Args:
-    json_data(dict): the loaded json data from results_overview
+    json_data(list): the loaded json data from results_overview
     name(str): the name of the dataset or the name of the approach
     key(str): the object that the function
                 should match on (either 'dataset' or 'recommender_system')
     """
     current_index = -1
     for i, data in enumerate(json_data):
-        result_value = json_data[i][key]
+        result_value = data[key]
         if by_name and result_value['name'] == name or result_value == name:
             current_index = i
     return current_index
