+++ server/tests/test_options_format.py
@@ -3,7 +3,7 @@ This program has been developed by students from the bachelor Computer Science a
 Utrecht University within the Software Project course.
 © Copyright Utrecht University (Department of Information and Computing Sciences)
 """
-from project.options_formatter import *
+from project.models.options_formatter import *
 
 form_options = {}
