+++ server/tests/test_result.py
@@ -4,10 +4,8 @@ Utrecht University within the Software Project course.
 © Copyright Utrecht University (Department of Information and Computing Sciences)
 """
 from unittest.mock import patch
-from project.result_storage import *
-from project.experiment import RecommenderSystem
-from tests.test_result_storage import test_experiment, test_id, save_mock_result, delete_test_results, \
-    test_results_path
+from project.models.result_storage import *
+from project.models.experiment import RecommenderSystem
 from tests.test_result_loader import MOCK_RESULTS_DIR
 
 url_prefix = '/api/result'
